add H2 and SQLite support and fixed bug.

+++ src/main/java/org/teasoft/exam/bee/test/ClearAndInitTestTable.java
@@ -8,8 +8,11 @@ package org.teasoft.exam.bee.test;
 
 import java.util.Map;
 
+import org.teasoft.bee.osql.DatabaseConst;
 import org.teasoft.bee.osql.PreparedSql;
+import org.teasoft.exam.bee.sqlite.CreateSQLiteTable;
 import org.teasoft.honey.osql.core.BeeFactory;
+import org.teasoft.honey.osql.core.HoneyConfig;
 import org.teasoft.honey.osql.core.Logger;
 
 /**
@@ -22,21 +25,28 @@ public class ClearAndInitTestTable {
 	public static void main(String[] args) {
 		PreparedSql preparedSql = BeeFactory.getHoneyFactory().getPreparedSql();
 		Map mapUpdate=null;
-//		int deleteOrdersNum=preparedSql.modify("delete from orders", mapUpdate);
-		int deleteOrdersNum=preparedSql.modify("TRUNCATE orders", mapUpdate);
-		Logger.info("deleteOrdersNum: "+deleteOrdersNum);
 		
-//		int deleteUserNum=preparedSql.modify("delete from user", mapUpdate);
-		int deleteUserNum=preparedSql.modify("TRUNCATE user", mapUpdate);
-		Logger.info("deleteUserNum: "+deleteUserNum);
+		String DbName=HoneyConfig.getHoneyConfig().getDbName();
 		
-		int deleteLeafNum=preparedSql.modify("TRUNCATE leaf_alloc", mapUpdate);
-		Logger.info("deleteLeafNum: "+deleteLeafNum);  //TRUNCATE 没有返回删除的行
+		if (DatabaseConst.SQLite.equalsIgnoreCase(DbName)) { //SQLite
+			CreateSQLiteTable.main(null);
+		} else {
+			//		int deleteOrdersNum=preparedSql.modify("delete from orders", mapUpdate);
+			int deleteOrdersNum = preparedSql.modify("TRUNCATE orders", mapUpdate);
+			Logger.info("deleteOrdersNum: " + deleteOrdersNum);
 
+			//		int deleteUserNum=preparedSql.modify("delete from user", mapUpdate);
+			int deleteUserNum = preparedSql.modify("TRUNCATE user", mapUpdate);
+			Logger.info("deleteUserNum: " + deleteUserNum);
+
+			int deleteLeafNum = preparedSql.modify("TRUNCATE leaf_alloc", mapUpdate);
+			Logger.info("deleteLeafNum: " + deleteLeafNum); //TRUNCATE 没有返回删除的行
+		}
 		
 	//>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>insert	
 		
-		String insertOrders1="INSERT INTO `orders` VALUES ('100001', 'bee', 'Bee(ORM Framework)', '95.01', '2020-03-02 11:29:28', 'test', '12345601', 'test1', '2020-03-02 16:34:19');";
+//		String insertOrders1="INSERT INTO `orders` VALUES ('100001', 'bee', 'Bee(ORM Framework)', '95.01', '2020-03-02 11:29:28', 'test', '12345601', 'test1', '2020-03-02 16:34:19');";
+		String insertOrders1="INSERT INTO `orders` VALUES ('100001', 'bee', 'Bee(ORM Framework)', '95.01', '2020-03-02 11:29:28', 'test', '12345601', 'test1', null);";
 		String insertOrders2="INSERT INTO `orders` VALUES ('100002', 'bee', 'Bee(ORM Framework)', '96.01', '2020-03-02 11:29:28', 'test%a', '12345602', 'test2', '2020-03-02 16:34:19');";
 		String insertOrders3="INSERT INTO `orders` VALUES ('100003', 'bee', 'Bee(ORM Framework)', '97.01', '2020-03-02 11:29:28', 'testx%', '12345603', 'test3', '2020-03-02 16:34:19');";
 		
@@ -61,8 +71,10 @@ public class ClearAndInitTestTable {
 		
 		String insertLeaf1="INSERT INTO `leaf_alloc` VALUES ('456537470402562', 'test', '100', '100', 'test segment', '2020-07-04 22:42:09', '1');";
 		String insertLeaf2="INSERT INTO `leaf_alloc` VALUES ('456537470402563', 'order', '100', '100', 'order segment', '2020-07-04 22:42:10', '1');";
+		String insertLeaf3="INSERT INTO `leaf_alloc` VALUES ('456537470402564', 'bee', '100', '100', 'order segment', '2020-09-19 19:05:10', '1');";
 		preparedSql.modify(insertLeaf1, mapUpdate);
 		preparedSql.modify(insertLeaf2, mapUpdate);
+		preparedSql.modify(insertLeaf3, mapUpdate);
 		
 	}
